requirements.txt
web3 == 7.8.0
python-dotenv == 1.0.0